move lipo to library
