+++ examples/test_battery_protector/BatteryProtectorCircuit.net
@@ -1,21 +1,21 @@
 (export (version D)
 (components
 (comp (ref "battery_protector.ic")
-  (value "S-8200ABE-M6T1U (ABLIC)")
+  (value "S-8261ABJMD-G3JT2S (ABLIC)")
   (footprint "Package_TO_SOT_SMD:SOT-23-6")
   (property (name "Sheetname") (value "battery_protector"))
-  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8200A.BatteryProtector_S8200A"))
+  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8261A.BatteryProtector_S8261A"))
   (property (name "edg_path") (value "battery_protector.ic"))
   (property (name "edg_short_path") (value "battery_protector.ic"))
   (property (name "edg_refdes") (value "U1"))
-  (property (name "edg_part") (value "S-8200ABE-M6T1U (ABLIC)"))
+  (property (name "edg_part") (value "S-8261ABJMD-G3JT2S (ABLIC)"))
   (sheetpath (names "/battery_protector/") (tstamps "/409b073d/"))
   (tstamps "013700cd"))
 (comp (ref "battery_protector.do_fet")
   (value "AO3400A (Alpha & Omega Semicon) - 30V 5.7A 26.5mΩ@10V,5.7A 1.4W 1.5V@250μA N Channel SOT-23-3L MOSFETs ROHS")
   (footprint "Package_TO_SOT_SMD:SOT-23")
   (property (name "Sheetname") (value "battery_protector"))
-  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8200A.BatteryProtector_S8200A"))
+  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8261A.BatteryProtector_S8261A"))
   (property (name "edg_path") (value "battery_protector.do_fet"))
   (property (name "edg_short_path") (value "battery_protector.do_fet"))
   (property (name "edg_refdes") (value "Q1"))
@@ -26,7 +26,7 @@
   (value "AO3400A (Alpha & Omega Semicon) - 30V 5.7A 26.5mΩ@10V,5.7A 1.4W 1.5V@250μA N Channel SOT-23-3L MOSFETs ROHS")
   (footprint "Package_TO_SOT_SMD:SOT-23")
   (property (name "Sheetname") (value "battery_protector"))
-  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8200A.BatteryProtector_S8200A"))
+  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8261A.BatteryProtector_S8261A"))
   (property (name "edg_path") (value "battery_protector.co_fet"))
   (property (name "edg_short_path") (value "battery_protector.co_fet"))
   (property (name "edg_refdes") (value "Q2"))
@@ -34,10 +34,10 @@
   (sheetpath (names "/battery_protector/") (tstamps "/409b073d/"))
   (tstamps "086f0271"))
 (comp (ref "battery_protector.vdd_res")
-  (value "330, 1%, 0.1 W")
+  (value "470, 1%, 0.1 W")
   (footprint "Resistor_SMD:R_0603_1608Metric")
   (property (name "Sheetname") (value "battery_protector"))
-  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8200A.BatteryProtector_S8200A"))
+  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8261A.BatteryProtector_S8261A"))
   (property (name "edg_path") (value "battery_protector.vdd_res.res"))
   (property (name "edg_short_path") (value "battery_protector.vdd_res"))
   (property (name "edg_refdes") (value "R1"))
@@ -48,7 +48,7 @@
   (value "100nF")
   (footprint "Capacitor_SMD:C_0603_1608Metric")
   (property (name "Sheetname") (value "battery_protector"))
-  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8200A.BatteryProtector_S8200A"))
+  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8261A.BatteryProtector_S8261A"))
   (property (name "edg_path") (value "battery_protector.vdd_vss_cap.cap"))
   (property (name "edg_short_path") (value "battery_protector.vdd_vss_cap"))
   (property (name "edg_refdes") (value "C1"))
@@ -59,7 +59,7 @@
   (value "2k, 1%, 0.1 W")
   (footprint "Resistor_SMD:R_0603_1608Metric")
   (property (name "Sheetname") (value "battery_protector"))
-  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8200A.BatteryProtector_S8200A"))
+  (property (name "Sheetfile") (value "electronics_lib.BatteryProtector_S8261A.BatteryProtector_S8261A"))
   (property (name "edg_path") (value "battery_protector.vm_res.res"))
   (property (name "edg_short_path") (value "battery_protector.vm_res"))
   (property (name "edg_refdes") (value "R2"))
